% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switchboard_helpers.R
\name{animate}
\alias{animate}
\title{easing function for timed animations}
\usage{
animate(theTime, timeMax, easer = "Bezier")
}
\description{
easing function for timed animations
}
\keyword{internal}
