% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{benchmark_construct}
\alias{benchmark_construct}
\title{helper function that constructs canvas items of a benchmark widget}
\usage{
benchmark_construct(.switchboard, label = " ", size = 1, ...)
}
\description{
helper function that constructs canvas items of a benchmark widget
}
\keyword{internal}
