% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{benchmark_update}
\alias{benchmark_update}
\title{helper function that updates canvas items of a benchmark widget}
\usage{
benchmark_update(.switchboard, eavesdrop, benchmark = NA, ...)
}
\description{
helper function that updates canvas items of a benchmark widget
}
\keyword{internal}
