% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caption.R
\name{caption_construct}
\alias{caption_construct}
\title{helper function that constructs canvas items of a caption widget}
\usage{
caption_construct(
  .switchboard,
  eavesdrop = c(NULL, NULL),
  extendRow = 1,
  size = 1,
  ...
)
}
\description{
helper function that constructs canvas items of a caption widget
}
\keyword{internal}
