% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control_slider_pair.R
\name{control_slider_pair_construct}
\alias{control_slider_pair_construct}
\title{helper function that constructs canvas items of a control_slider_pair widget}
\usage{
control_slider_pair_construct(
  .switchboard,
  inject = c("", ""),
  minimum = c(0, 0),
  maximum = c(100, 100),
  label = c(" ", " "),
  size = 1,
  ...
)
}
\description{
helper function that constructs canvas items of a control_slider_pair widget
}
\keyword{internal}
