% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control_slider_Y.R
\name{control_slider_Y_construct}
\alias{control_slider_Y_construct}
\title{helper function that constructs canvas items of a control_slider_Y widget}
\usage{
control_slider_Y_construct(
  .switchboard,
  inject = "",
  minimum = 0,
  maximum = 100,
  label = " ",
  size = 1,
  ...
)
}
\description{
helper function that constructs canvas items of a control_slider_Y widget
}
\keyword{internal}
