% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control_switch.R
\name{control_switch_construct}
\alias{control_switch_construct}
\title{helper function that constructs canvas items of a control_switch widget}
\usage{
control_switch_construct(.switchboard, inject = "", label = " ", size = 1, ...)
}
\description{
helper function that constructs canvas items of a control_switch widget
}
\keyword{internal}
