% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control_switch_pair.R
\name{control_switch_pair_construct}
\alias{control_switch_pair_construct}
\title{helper function that constructs canvas items of a control_switch_pair widget}
\usage{
control_switch_pair_construct(
  .switchboard,
  inject = c("", ""),
  label = c("", ""),
  size = 1,
  ...
)
}
\description{
helper function that constructs canvas items of a control_switch_pair widget
}
\keyword{internal}
