% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control_switch_trio_R.R
\name{control_switch_trio_construct}
\alias{control_switch_trio_construct}
\title{helper function that constructs canvas items of a control_switch_trio widget}
\usage{
control_switch_trio_construct(
  .switchboard,
  inject = c("", "", ""),
  label = c("", "", ""),
  size = 1,
  ...
)
}
\description{
helper function that constructs canvas items of a control_switch_trio widget
}
\keyword{internal}
