% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counter.R
\name{counter_construct}
\alias{counter_construct}
\title{helper function that constructs canvas items of a counter widget}
\usage{
counter_construct(.switchboard, label = " ", size = 1, ...)
}
\description{
helper function that constructs canvas items of a counter widget
}
\keyword{internal}
