% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counter_tally.R
\name{counter_tally_construct}
\alias{counter_tally_construct}
\title{helper function that constructs canvas items of a counter_tally widget}
\usage{
counter_tally_construct(.switchboard, label = " ", size = 1, ...)
}
\description{
helper function that constructs canvas items of a counter_tally widget
}
\keyword{internal}
