% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counter_tally.R
\name{counter_tally_update}
\alias{counter_tally_update}
\title{helper function that updates canvas items of a counter_tally widget}
\usage{
counter_tally_update(.switchboard, ...)
}
\description{
helper function that updates canvas items of a counter_tally widget
}
\keyword{internal}
