% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counter.R
\name{counter_update}
\alias{counter_update}
\title{helper function that updates canvas items of a counter widget}
\usage{
counter_update(.switchboard, ...)
}
\description{
helper function that updates canvas items of a counter widget
}
\keyword{internal}
