% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eavesdropper_2D.R
\name{eavesdropper_2D_construct}
\alias{eavesdropper_2D_construct}
\title{helper function that constructs canvas items of a eavesdropper_2D widget}
\usage{
eavesdropper_2D_construct(
  .switchboard,
  eavesdrop = c(NULL, NULL),
  minimum = c(" ", " "),
  maximum = c(" ", " "),
  label = c(" ", " "),
  size = 1,
  switch = FALSE,
  ...
)
}
\description{
helper function that constructs canvas items of a eavesdropper_2D widget
}
\keyword{internal}
