% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eavesdropper.R
\name{eavesdropper_construct}
\alias{eavesdropper_construct}
\title{helper function that constructs canvas items of a eavesdropper widget}
\usage{
eavesdropper_construct(
  .switchboard,
  label = " ",
  size = 1,
  switch = FALSE,
  ...
)
}
\description{
helper function that constructs canvas items of a eavesdropper widget
}
\keyword{internal}
