% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eavesdropper_X.R
\name{eavesdropper_X_construct}
\alias{eavesdropper_X_construct}
\title{helper function that constructs canvas items of a eavesdropper_X widget}
\usage{
eavesdropper_X_construct(
  .switchboard,
  label = " ",
  size = 1,
  switch = FALSE,
  ...
)
}
\description{
helper function that constructs canvas items of a eavesdropper_X widget
}
\keyword{internal}
