% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eavesdropper_X.R
\name{eavesdropper_X_update}
\alias{eavesdropper_X_update}
\title{helper function that updates the canvas items of a eavesdropper_X widget}
\usage{
eavesdropper_X_update(
  .switchboard,
  eavesdrop = NULL,
  minimum = " ",
  maximum = " ",
  size = 1,
  updates = 1,
  plotMean = FALSE,
  ...
)
}
\description{
helper function that updates the canvas items of a eavesdropper_X widget
}
\keyword{internal}
