% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/injector_2D.R
\name{injector_2D_construct}
\alias{injector_2D_construct}
\title{helper function that constructs canvas items of a injector_2D widget}
\usage{
injector_2D_construct(
  .switchboard,
  eavesdrop = c(NULL, NULL),
  inject = c("", ""),
  minimum = c(" ", " "),
  maximum = c(" ", " "),
  label = c(" ", " "),
  size = 1,
  switch = FALSE,
  ...
)
}
\description{
helper function that constructs canvas items of a injector_2D widget
}
\keyword{internal}
