% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/injector_2D.R
\name{injector_2D_update}
\alias{injector_2D_update}
\title{helper function that updates the canvas items of a injector_2D widget}
\usage{
injector_2D_update(
  .switchboard,
  eavesdrop = c(NULL, NULL),
  minimum = c(" ", " "),
  maximum = c(" ", " "),
  size = 1,
  updates = 1,
  forget = 400,
  plotRegression = FALSE,
  plotSampleSize = FALSE,
  ...
)
}
\description{
helper function that updates the canvas items of a injector_2D widget
}
\keyword{internal}
