% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/injector.R
\name{injector_construct}
\alias{injector_construct}
\title{helper function that constructs canvas items of a injector widget}
\usage{
injector_construct(
  .switchboard,
  label = " ",
  minimum = " ",
  maximum = " ",
  inject = "",
  size = 1,
  switch = FALSE,
  ...
)
}
\description{
helper function that constructs canvas items of a injector widget
}
\keyword{internal}
