% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/injector.R
\name{injector_update}
\alias{injector_update}
\title{helper function that updates the canvas items of a injector widget}
\usage{
injector_update(
  .switchboard,
  eavesdrop = NULL,
  minimum = " ",
  maximum = " ",
  size = 1,
  updates = 1,
  plotMean = FALSE,
  ...
)
}
\description{
helper function that updates the canvas items of a injector widget
}
\keyword{internal}
