% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/injector_X.R
\name{injector_X_construct}
\alias{injector_X_construct}
\title{helper function that constructs canvas items of a injector_X widget}
\usage{
injector_X_construct(
  .switchboard,
  label = " ",
  inject = "",
  minimum = " ",
  maximum = " ",
  switch = FALSE,
  size = 1,
  ...
)
}
\description{
helper function that constructs canvas items of a injector_X widget
}
\keyword{internal}
