% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/injector_X.R
\name{injector_X_update}
\alias{injector_X_update}
\title{helper function that updates the canvas items of a injector_X widget}
\usage{
injector_X_update(
  .switchboard,
  eavesdrop = NULL,
  minimum = " ",
  maximum = " ",
  size = 1,
  updates = 1,
  plotMean = FALSE,
  ...
)
}
\description{
helper function that updates the canvas items of a injector_X widget
}
\keyword{internal}
