% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number.R
\name{number_construct}
\alias{number_construct}
\title{helper function that constructs canvas items of a number widget}
\usage{
number_construct(.switchboard, eavesdrop = NULL, size = 1, label = " ", ...)
}
\description{
helper function that constructs canvas items of a number widget
}
\keyword{internal}
