% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number_pair.R
\name{number_pair_construct}
\alias{number_pair_construct}
\title{helper function that constructs canvas items of a number_pair widget}
\usage{
number_pair_construct(
  .switchboard,
  eavesdrop = c(NULL, NULL),
  size = 1,
  label = c(" ", " "),
  ...
)
}
\description{
helper function that constructs canvas items of a number_pair widget
}
\keyword{internal}
