% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number_pair.R
\name{number_pair_update}
\alias{number_pair_update}
\title{helper function that updates canvas items of a number_pair widget}
\usage{
number_pair_update(
  .switchboard,
  eavesdrop = c(NULL, NULL),
  digits = 5,
  updates = 1,
  ...
)
}
\description{
helper function that updates canvas items of a number_pair widget
}
\keyword{internal}
