% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number_quartet.R
\name{number_quartet_construct}
\alias{number_quartet_construct}
\title{helper function that constructs canvas items of a number_quartet widget}
\usage{
number_quartet_construct(
  .switchboard,
  eavesdrop = c(NULL, NULL, NULL, NULL),
  size = 1,
  digits = 5,
  label = c(" ", " ", " ", " "),
  ...
)
}
\description{
helper function that constructs canvas items of a number_quartet widget
}
\keyword{internal}
