% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number_quartet.R
\name{number_quartet_update}
\alias{number_quartet_update}
\title{helper function that updates canvas items of a number_quartet widget}
\usage{
number_quartet_update(
  .switchboard,
  eavesdrop = c(NULL, NULL, NULL),
  digits = 5,
  updates = 1,
  ...
)
}
\description{
helper function that updates canvas items of a number_quartet widget
}
\keyword{internal}
