% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number_trio.R
\name{number_trio}
\alias{number_trio}
\title{Display the value of three continuous variables in a simulation.}
\arguments{
\item{eavesdrop}{A vector of three variables to track.}

\item{digits}{The number of digits to display in a widget.}

\item{label}{A vector of three small captions/labels for the widget.}

\item{updates}{The number of times the widget is to be updated (e.g., when
it be modified/changed). The default updates with each iteration.}

\item{size}{A number used to designate the size (magnification) of the
widget. The default is set to \code{1} which is 80 by 80 pixels. For
example, setting to \code{3} will results in a widget 3-times the default
size (240 by 240 pixels) and will occupy a grid area of 3 by 3.}

\item{placeOnGrid}{A row by column coordinate (e.g., \code{c(row-number, column-number)})
of a grid that designates the position to draw the widget on the
\code{switchboard}. Use \code{showGrid()} to help organize widget placement
on dashboard. The default places the first widget in pipe chain to
the \code{c(1, 1)} position, and all following on the same row stacked to
the right.}
}
\value{
Nothing.
}
\description{
The \code{number_trio} widget displays the numerical value of three continuous
variables.
}
\section{Usage}{

   \preformatted{number_trio(eavesdrop = c(NULL, NULL, NULL), digits = 5,
   updates = 1, label = c("", "", ""), size = 1, placeOnGrid = c(1, 1))}
}

\examples{
\dontrun{
     
     for (i in 1:250) {
       switchboard(delay = 0.01) \%>\%
         number_trio(c(i, i/2, i/3), label = c("i", "i/2", "i/3"))
     }
     switchboard_close()

}

}
\seealso{
Other eavesdroppers: 
\code{\link{benchmark}()},
\code{\link{counter_tally}()},
\code{\link{counter}()},
\code{\link{eavesdropper_2D}()},
\code{\link{eavesdropper_X}()},
\code{\link{eavesdropper}()},
\code{\link{injector_2D}()},
\code{\link{injector_X}()},
\code{\link{injector}()},
\code{\link{number_pair}()},
\code{\link{number_quartet}()},
\code{\link{number}()},
\code{\link{progress_benchmark}()}
}
\concept{eavesdroppers}
