% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number_trio.R
\name{number_trio_construct}
\alias{number_trio_construct}
\title{helper function that constructs canvas items of a number_trio widget}
\usage{
number_trio_construct(
  .switchboard,
  eavesdrop = c(NULL, NULL, NULL),
  size = 1,
  digits = 5,
  label = c(" ", " ", " "),
  ...
)
}
\description{
helper function that constructs canvas items of a number_trio widget
}
\keyword{internal}
