% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number_trio.R
\name{number_trio_update}
\alias{number_trio_update}
\title{helper function that updates canvas items of a number_trio widget}
\usage{
number_trio_update(
  .switchboard,
  eavesdrop = c(NULL, NULL, NULL),
  digits = 5,
  updates = 1,
  ...
)
}
\description{
helper function that updates canvas items of a number_trio widget
}
\keyword{internal}
