% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number.R
\name{number_update}
\alias{number_update}
\title{helper function that updates canvas items of a number widget}
\usage{
number_update(.switchboard, eavesdrop = NULL, digits = 5, updates = 1, ...)
}
\description{
helper function that updates canvas items of a number widget
}
\keyword{internal}
