% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress_benchmark.R
\name{progress_benchmark_construct}
\alias{progress_benchmark_construct}
\title{helper function that constructs canvas items of a progress_benchmark widget}
\usage{
progress_benchmark_construct(.switchboard, size = 1, ...)
}
\description{
helper function that constructs canvas items of a progress_benchmark widget
}
\keyword{internal}
