% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress_benchmark.R
\name{progress_benchmark_update}
\alias{progress_benchmark_update}
\title{helper function that updates the canvas items of a progress_benchmark widget}
\usage{
progress_benchmark_update(
  .switchboard,
  eavesdrop = c(NULL, NULL),
  maximum = c(" ", " "),
  updates = 100,
  honest = FALSE,
  closeAtMaximum = FALSE,
  benchmark = NA,
  ...
)
}
\description{
helper function that updates the canvas items of a progress_benchmark widget
}
\keyword{internal}
