% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress_image.R
\name{progress_image_construct}
\alias{progress_image_construct}
\title{helper function that constructs canvas items of a progress_image widget}
\usage{
progress_image_construct(
  .switchboard,
  size = 1,
  file = system.file("images", "topography.png", package = "switchboard"),
  ...
)
}
\description{
helper function that constructs canvas items of a progress_image widget
}
\keyword{internal}
