% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress_image.R
\name{progress_image_update}
\alias{progress_image_update}
\title{helper function that updates the canvas items of a progress_image widget}
\usage{
progress_image_update(
  .switchboard,
  eavesdrop = NULL,
  maximum = 100,
  size = 1,
  updates = maximum,
  honest = TRUE,
  closeAtMaximum = FALSE,
  fill = "horizontal",
  ...
)
}
\description{
helper function that updates the canvas items of a progress_image widget
}
\keyword{internal}
