% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress_percent.R
\name{progress_percent_construct}
\alias{progress_percent_construct}
\title{helper function that constructs canvas items of a progress_image widget}
\usage{
progress_percent_construct(.switchboard, label = " ", size = 1, ...)
}
\description{
helper function that constructs canvas items of a progress_image widget
}
\keyword{internal}
