% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress_percent.R
\name{progress_percent_update}
\alias{progress_percent_update}
\title{helper function that updates the canvas items of a progress_image widget}
\usage{
progress_percent_update(
  .switchboard,
  eavesdrop = NULL,
  maximum = 100,
  size = 1,
  updates = 100,
  honest = FALSE,
  closeAtMaximum = FALSE,
  ...
)
}
\description{
helper function that updates the canvas items of a progress_image widget
}
\keyword{internal}
