% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress_phyllotaxis.R
\name{progress_phyllotaxis_construct}
\alias{progress_phyllotaxis_construct}
\title{helper function that constructs canvas items of a progress_phyllotaxis widget}
\usage{
progress_phyllotaxis_construct(.switchboard, label = " ", size = 1, ...)
}
\description{
helper function that constructs canvas items of a progress_phyllotaxis widget
}
\keyword{internal}
