% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress_phyllotaxis.R
\name{progress_phyllotaxis_update}
\alias{progress_phyllotaxis_update}
\title{helper function that updates the canvas items of a progress_phyllotaxis widget}
\usage{
progress_phyllotaxis_update(
  .switchboard,
  eavesdrop,
  maximum = 100,
  size = 1,
  updates = 100,
  closeAtMaximum = FALSE,
  ...
)
}
\description{
helper function that updates the canvas items of a progress_phyllotaxis widget
}
\keyword{internal}
