% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress_ring.R
\name{progress_ring_construct}
\alias{progress_ring_construct}
\title{helper function that constructs canvas items of a progress_ring widget}
\usage{
progress_ring_construct(.switchboard, label = " ", size = 1, ...)
}
\description{
helper function that constructs canvas items of a progress_ring widget
}
\keyword{internal}
