% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress_ring_percent.R
\name{progress_ring_percent_construct}
\alias{progress_ring_percent_construct}
\title{helper function that constructs canvas items of a progress_ring_percent widget}
\usage{
progress_ring_percent_construct(.switchboard, label = " ", size = 1, ...)
}
\description{
helper function that constructs canvas items of a progress_ring_percent widget
}
\keyword{internal}
