% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress_ring.R
\name{progress_ring_update}
\alias{progress_ring_update}
\title{helper function that updates the canvas items of a progress_ring widget}
\usage{
progress_ring_update(
  .switchboard,
  eavesdrop,
  maximum = 100,
  size = 1,
  updates = 100,
  honest = FALSE,
  closeAtMaximum = FALSE,
  ...
)
}
\description{
helper function that updates the canvas items of a progress_ring widget
}
\keyword{internal}
