% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switchboard_package.R
\docType{package}
\name{switchboard-package}
\alias{switchboard-package}
\title{Agile widgets for visualizing and interacting with simulations in R.}
\description{
\pkg{switchboard} is an experimental package for generating agile dashboards
for iterative simulations in R. The switchboard window is rendered in 'Tcl/Tk'
\url{https://www.tcl.tk/} which is the GUI (graphical user interface) toolkit
of base R.  More information about \pkg{switchboard} can be found 
at \url{http://lajeunesse.myweb.usf.edu/}.
}
\author{
Marc J. Lajeunesse (University of South Florida, Tampa USA)
}
