% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switchboard_close.R
\name{switchboard_close}
\alias{switchboard_close}
\title{Closes all switchboard windows.}
\usage{
switchboard_close()
}
\value{
Nothing.
}
\description{
Helps remove all visible/invisible switchboard windows. Place outside of 
loop to remove windows.
}
\examples{
\dontrun{

for(i in 1:100) 
    switchboard(delay = 0.05) \%>\% percent(i, maximum = 100, label = "100 loops")
switchboard_close()

}

}
\seealso{
Other dashboard: 
\code{\link{showGrid}()},
\code{\link{switchboard}()}
}
\concept{dashboard}
