% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switchboard_helpers.R
\name{value_to_coordinate}
\alias{value_to_coordinate}
\title{converting raw values into standardized coordinates within an 80 by 80 widget}
\usage{
value_to_coordinate(theValue, min_value, max_value, size = 1, adjust = c(0, 0))
}
\description{
converting raw values into standardized coordinates within an 80 by 80 widget
}
\keyword{internal}
